feature(cart): ajaxify cart.
This commit moves the cart to the side bar of the application and
makes it disappear if the cart is empty.

It also adds ActionCable capabilities to help with reloading
store content if they are changed from another source.

The cart autoreloads to notify the user if something has been
updated.